# [Arm64] Treat Math/MathF.FusedMultiplyAdd as intrinsics