@echesakovMSFT 2020-08-06
@dotnet/jit-contrib @tannergooding I think this should be ready for review now. 

There is no codegen changes across all frameworks libraries since there no uses of `MathF.FusedMultiplyAdd` or `Math.FusedMultiplyAdd`.