@echesakovMSFT 2020-08-06
/azp run runtime-live-build